requirements.txt
Django==4.0.1
django-formtools==2.3
django-crispy-forms==1.14.0
django-tempus-dominus==5.1.2
django-tables2=2.4.1